+++ Assets/Scripts/PlayerScripts/PlayerMovement.cs
@@ -1,0 +1,45 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerMovement : MonoBehaviour
+{
+    // Controller Reference
+    public CharacterController controller;
+    public Transform GroundCheck;
+    public Camera ScientistCamera;
+
+    // Values
+    public float speed = 3f;
+    public float gravity = -9.81f;
+    public float GroundDistance = 0.4f;
+    public LayerMask GroundMask;
+
+    // Velocity
+    Vector3 velocity;
+    bool isGrounded;
+
+    // Update is called once per frame
+    void Update()
+    {
+        isGrounded = Physics.CheckSphere(GroundCheck.position, GroundDistance, GroundMask);
+
+        if (isGrounded && velocity.y < 0)
+        {
+            velocity.y = -2f;
+        }
+        if (ScientistCamera.enabled)
+        {
+            float x = Input.GetAxis("Horizontal");
+            float z = Input.GetAxis("Vertical");
+
+            Vector3 move = transform.right * x + transform.forward * z;
+
+            controller.Move(move * speed * Time.deltaTime);
+        }        
+
+        velocity.y += gravity * Time.deltaTime;
+
+        controller.Move(velocity * Time.deltaTime);
+    }
+}